Reorganize FE Common components
