+++ frontend/src/components/Common/NavigationCard/NavigationCard.module.css
