+++ frontend/src/pages/admin/index.tsx
@@ -1,4 +1,4 @@
-import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard';
+import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard/NavigationCard';
 
 const navCardItems: readonly NavigationCardItem[] = [
   {
