+++ frontend/src/pages/games/index.tsx
@@ -1,4 +1,4 @@
-import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard';
+import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard/NavigationCard';
 
 const navCardItems: readonly NavigationCardItem[] = [
   { header: 'DTI48', description: 'Keep merging until you get Ashneel.', link: '/games/dti48' }
